requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==0.4.3
altair==4.2.0
antlr4-python3-runtime==4.8
anyio==3.6.1
async-timeout==4.0.2
attrs==20.3.0
autopep8==1.7.0
backports.zoneinfo==0.2.1
blinker==1.5
brotlipy==0.7.0
cachetools==5.2.0
cbor2==5.4.3
click==8.1.3
-e git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1#egg=clip
coloredlogs==15.0.1
commonmark==0.9.1
decorator==5.1.1
diffusers==0.3.0
einops==0.3.0
entrypoints==0.4
fastapi==0.82.0
filelock==3.8.0
flatbuffers==2.0.7
frozenlist==1.3.1
fsspec==2022.8.2
ftfy==6.1.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.48.1
h11==0.13.0
httptools==0.4.0
huggingface-hub==0.9.1
humanfriendly==10.0
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.2.6
importlib-metadata==4.12.0
importlib-resources==5.9.0
invisible-watermark==0.1.5
Jinja2==3.1.2
jsonschema==4.15.0
kornia==0.6.0
-e git+https://github.com/CompVis/stable-diffusion.git@69ae4b35e0a0f6ee1af8bb9a5d0016ccb27e36dc#egg=latent_diffusion
Markdown==3.4.1
MarkupSafe==2.1.1
mkl-fft==1.3.1
mkl-service==2.4.0
mpmath==1.2.1
multidict==6.0.2
networkx==2.8.6
numpy==1.23.2
oauthlib==3.2.0
omegaconf==2.1.1
onnx==1.12.0
onnxruntime==1.12.1
opencv-python==4.1.2.30
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
pkgutil-resolve-name==1.3.10
protobuf==3.19.4
pudb==2019.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pydantic==1.10.2
pydeck==0.8.0b1
pyDeprecate==0.3.1
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytorch-lightning==1.4.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0
regex==2022.8.17
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
scikit-image==0.19.3
scipy==1.9.1
semver==2.13.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sniffio==1.3.0
starlette==0.19.1
streamlit==1.12.2
stringcase==1.2.0
structlog==20.2.0
sympy==1.11.1
-e git+https://github.com/CompVis/taming-transformers.git@24268930bf1dce879235a7fddd0b2355b84d7ea6#egg=taming_transformers
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
test-tube==0.7.5
tifffile==2022.8.12
tokenizers==0.12.1
toml==0.10.2
toolz==0.12.0
torch==1.11.0
torch-fidelity==0.3.0
torchmetrics==0.6.0
torchvision==0.12.0
tornado==6.2
tqdm==4.64.1
transformers==4.19.2
tzdata==2022.2
tzlocal==4.2
urwid==2.1.2
uvicorn==0.18.3
uvloop==0.16.0
validators==0.20.0
watchdog==2.1.9
watchfiles==0.16.1
wcwidth==0.2.5
websockets==10.3
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.8.1
